+++ src/java/org/infoglue/cms/applications/contenttool/actions/UpdateDigitalAssetAction.java
@@ -125,7 +125,8 @@ public class UpdateDigitalAssetAction extends ViewDigitalAssetAction
 						
 						String fileUploadMaximumSize = getPrincipalPropertyValue("fileUploadMaximumSize", false, true);
 						System.out.println("fileUploadMaximumSize:" + fileUploadMaximumSize);
-						if(new Integer(fileUploadMaximumSize).intValue() < new Long(file.length()).intValue())
+						
+						if(!fileUploadMaximumSize.equalsIgnoreCase("-1") && new Integer(fileUploadMaximumSize).intValue() < new Long(file.length()).intValue())
 						{
 						    file.delete();
 						    this.reasonKey = "tool.contenttool.fileUpload.fileUploadFailedOnSizeText";
